Complete changing transport to RabbitMQ
I thought the previous commit had changed my solution to use RabbitMQ
transport; however, my "test" only verified that the web page correctly
displayed the correct message "text" (and ensured that no exception was
thrown). The test **did not** actually verify that the `UserService` and
`WelcomeEmailService` end points **received** and handled the message
correctly.

This commit corrects that error by configuring the `ExampleWeb` and
`WelcomeEmailService` host endpoints to use the RabbitMQ transport.

In addition, this commit replaces the scripts to clean up only the
`UserService` exchanges and queues with a script that cleans up all
three created exchanges and queues:

* `ExampleWeb`
* `UserService`
* `WelcomeEmailService`

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch rabbitmq
# Changes to be committed:
#	modified:   Example/ExampleWeb/ExampleWeb.csproj
#	modified:   Example/ExampleWeb/ServiceBus.cs
#	modified:   Example/ExampleWeb/Web.config
#	modified:   Example/ExampleWeb/packages.config
#	modified:   Example/UserService/UserCreator.cs
#	modified:   Example/UserService/UserService.csproj
#	modified:   Example/UserService/packages.config
#	modified:   Example/WelcomeEmailService/App.config
#	modified:   Example/WelcomeEmailService/EndpointConfig.cs
#	modified:   Example/WelcomeEmailService/WelcomeEmailService.csproj
#	modified:   Example/WelcomeEmailService/packages.config
#	deleted:    delete_user_service_exchanges.sh
#	deleted:    delete_user_service_queues.sh
#	new file:   rabbit_clean_up.sh
#