Add feature to display recent subscribers
Added a feature to display recent subscribers. The `ExampleWeb` project
now displays the "subscriber" list at the index page. This list is
updated by the `ExampleWeb` project whenever it processes an
`IUserCreatedEvent`.

The solution does not currently redirect to the index page, but after a
manual refresh, it correctly displays the updated subscriber list.